drm/panel: shift-sh8804b: new panel
Add a driver for the SH8804B panel found in the SHIFTphone 8. The panel
doesn't seem to have a vendor? And it's likely to be highly specific to
this device anyways.

The panel can do 120hz however the current init sequence and
configuration is fixed to 60hz.

[luca.weiss@fairphone.com: regen from dtb]
Signed-off-by: Casey Connolly <casey.connolly@linaro.org>

NOTE: This is generated from downstream 'sh8804b_mz_amoled_fhd_plus_cmd'

TODO: Shall we use sh8804b_mz_amoled_dsc_fhd_plus_cmd? But that makes
more artifacts, not sure if brightness would be any better...